server/requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ipython
psycopg2-binary==2.9.5